+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this extension will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+* Setting to hide CPU usage
+* Setting to hide RAM usage
+
 ## [0.1.0 - 2023-02-21]
 
 ### Added